minor - introduce eq univariate, move tests and bit_reverse to utils, fix one dumb computation
